project/requirements.txt
pygments>=2.0,<2.1
simplejson==3.13.2
python-dateutil>=2.4,<2.5
requests==2.21.0
sqlparse==0.2.4
jinja2==2.10
autopep8==1.4.3
pytz>2014.2
mock==2.0.0
pillow==5.4.1
factory-boy==2.9.2
freezegun==0.3.12
networkx==1.11
pydotplus==2.0.2
contextlib2==0.5.5
coverage==5.0.3
pytest-django==3.8.0
pytest-cov==2.8.1
